Updated the Arduino code to properly transmit the readings according to the GATT specifications. Added temperature and humidity to ConnectedScreen.

+++ app/src/main/java/com/example/umco2monitor/SensorViewModel.kt
@@ -67,12 +67,12 @@ class SensorViewModel(private val application: Application) : ViewModel() {
     /**
      * The most recent temperature value, sourced from BluetoothHandler.
      */
-    val temperatureValue: StateFlow<Short?> = BluetoothHandler.temperatureValue
+    val temperatureValue: StateFlow<Float?> = BluetoothHandler.temperatureValue
 
     /**
      * The most recent humidity value, sourced from BluetoothHandler.
      */
-    val humidityValue: StateFlow<UShort?> = BluetoothHandler.humidityValue
+    val humidityValue: StateFlow<Float?> = BluetoothHandler.humidityValue
 
     /**
      * The most recent battery level value, sourced from BluetoothHandler.
